@@ -1,32 +1,32 @@
 {
     "name": "node-dashboard",
     "description": "Dashboard for Nodes",
-    "version": "1.0.0",
+    "version": "2.0.0-rc.1",
     "author": "Martyn Janes <martyn.janes@iota.org>",
     "repository": {
         "type": "git",
         "url": "https://github.com/iotaledger/node-dashboard"
     },
+    "homepage": "/dashboard",
     "license": "MIT",
     "dependencies": {
-        "@iota/iota.js": "^1.5.1",
+        "@iota/iota.js": "^2.0.0-rc.1",
         "classnames": "^2.3.1",
-        "humanize-duration": "^3.25.2",
+        "humanize-duration": "^3.27.3",
         "moment": "^2.29.1",
-        "react": "^17.0.2",
-        "react-dom": "^17.0.2",
+        "react": "^18.2.0",
+        "react-dom": "^18.2.0",
         "react-router-dom": "^5.2.0",
-        "react-scripts": "^4.0.3",
         "vivagraphjs": "^0.12.0"
     },
     "engines": {
-        "node": "14.x.x"
+        "node": ">=14 <=16.16"
     },
     "scripts": {
         "start": "cross-env PORT=9090 WDS_SOCKET_PORT=9090 REACT_APP_SOCKET=8081 REACT_APP_BRAND_ID=hornet craco start",
         "start-bee": "cross-env PORT=9090 WDS_SOCKET_PORT=9090 REACT_APP_SOCKET=8081 REACT_APP_BRAND_ID=bee craco start",
         "lint": "eslint src --ext .tsx,.ts",
-        "sass-lint": "stylelint ./src/**/*.scss",
+        "sass-lint": "stylelint ./src/**/*.scss --custom-syntax postcss-scss",
         "build-internal": "craco build",
         "build": "cross-env REACT_APP_BRAND_ID=hornet run-s lint sass-lint build-internal",
         "build-bee": "cross-env REACT_APP_BRAND_ID=bee run-s lint sass-lint build-internal",
@@ -40,31 +40,37 @@
         "not op_mini all"
     ],
     "devDependencies": {
-        "@craco/craco": "^6.1.2",
+        "@craco/craco": "^7.0.0-alpha.7",
         "@types/classnames": "^2.3.1",
         "@types/humanize-duration": "^3.25.0",
-        "@types/node": "^15.3.0",
-        "@types/react": "^17.0.5",
-        "@types/react-dom": "^17.0.5",
+        "@types/node": "^18.7.18",
+        "@types/react": "^18.0.21",
+        "@types/react-dom": "^18.0.6",
         "@types/react-helmet": "^6.1.1",
         "@types/react-router-dom": "^5.1.7",
-        "@typescript-eslint/eslint-plugin": "^4.23.0",
-        "@typescript-eslint/parser": "^4.23.0",
+        "@typescript-eslint/eslint-plugin": "^5.38.0",
+        "@typescript-eslint/parser": "^5.30.7",
+        "buffer": "^6.0.3",
         "cross-env": "^7.0.3",
-        "eslint": "^7.26.0",
+        "eslint": "^8.20.0",
         "eslint-plugin-import": "^2.23.2",
-        "eslint-plugin-jsdoc": "^34.6.3",
-        "eslint-plugin-react": "^7.23.2",
-        "eslint-plugin-unicorn": "^32.0.1",
+        "eslint-plugin-jsdoc": "^39.3.3",
+        "eslint-plugin-react": "^7.30.1",
+        "eslint-plugin-unicorn": "^43.0.2",
         "npm-run-all": "^4.1.5",
-        "sass": "^1.32.13",
-        "stylelint": "^13.13.0",
-        "stylelint-config-property-sort-order-smacss": "^7.1.0",
-        "stylelint-config-recommended": "^5.0.0",
-        "stylelint-config-recommended-scss": "^4.2.0",
-        "stylelint-config-sass-guidelines": "^8.0.0",
-        "stylelint-config-standard": "^22.0.0",
-        "stylelint-scss": "^3.19.0",
-        "typescript": "^4.2.4"
+        "postcss-scss": "^4.0.5",
+        "react-scripts": "^5.0.1",
+        "sass": "^1.55.0",
+        "stylelint": "^14.12.1",
+        "stylelint-config-property-sort-order-smacss": "^9.0.0",
+        "stylelint-config-recommended": "^9.0.0",
+        "stylelint-config-recommended-scss": "7.0.0",
+        "stylelint-config-sass-guidelines": "^9.0.1",
+        "stylelint-config-standard": "^28.0.0",
+        "stylelint-scss": "^4.0.0",
+        "typescript": "^4.7.4"
+    },
+    "overrides": {
+        "@svgr/webpack": "^6.3.1"
     }
 }
