@@ -1,5 +1,6 @@
+/* eslint-disable unicorn/prefer-top-level-await */
 import React from "react";
-import ReactDOM from "react-dom";
+import { createRoot } from "react-dom/client";
 import { BrowserRouter } from "react-router-dom";
 import App from "./app/App";
 import Participation from "./app/components/plugins/Participation";
@@ -12,6 +13,7 @@ import { EventAggregator } from "./services/eventAggregator";
 import { LocalStorageService } from "./services/localStorageService";
 import { MetricsService } from "./services/metricsService";
 import { NodeConfigService } from "./services/nodeConfigService";
+import { SessionStorageService } from "./services/sessionStorageService";
 import { SettingsService } from "./services/settingsService";
 import { TangleService } from "./services/tangleService";
 import { ThemeService } from "./services/themeService";
@@ -21,15 +23,17 @@ import { BrandHelper } from "./utils/brandHelper";
 
 initServices()
     .then(brandConfiguration => {
-        ReactDOM.render(
+        /* eslint-disable-next-line @typescript-eslint/no-non-null-assertion */
+        const container = document.querySelector("#root")!;
+        const root = createRoot(container);
+        root.render(
             !brandConfiguration
                 ? (<div>REACT_APP_BRAND_ID is not set</div>)
                 : (
-                    <BrowserRouter>
+                    <BrowserRouter basename={process.env.PUBLIC_URL}>
                         <App />
                     </BrowserRouter>
-                ),
-            document.querySelector("#root")
+                )
         );
     })
     .catch(err => console.error(err));
@@ -39,7 +43,8 @@ initServices()
  * @returns The brand configuration.
  */
 async function initServices(): Promise<IBrandConfiguration | undefined> {
-    ServiceFactory.register("storage", () => new LocalStorageService());
+    ServiceFactory.register("local-storage", () => new LocalStorageService());
+    ServiceFactory.register("session-storage", () => new SessionStorageService());
     const settingsService = new SettingsService();
     ServiceFactory.register("settings", () => settingsService);
 
