@@ -1,18 +1,28 @@
-import { IOutputResponse } from "@iota/iota.js";
+import { IOutputMetadataResponse, OutputTypes } from "@iota/iota.js";
 
 export interface OutputProps {
     /**
-     * The output id.
-     */
-    id: string;
-
-    /**
      * The index within the parent.
      */
-    index: number;
+    index?: number;
 
     /**
-     * The output to display.
+     * The output metadata.
      */
-    output: IOutputResponse;
+    metadata?: IOutputMetadataResponse;
+
+    /**
+     * The output id.
+     */
+    outputId: string;
+
+    /**
+     * The output.
+     */
+    output: OutputTypes;
+
+    /**
+     * Expand output details.
+     */
+    showDetails?: boolean;
 }
