@@ -1,4 +1,5 @@
-import { Converter, INDEXATION_PAYLOAD_TYPE, MILESTONE_PAYLOAD_TYPE, SIG_LOCKED_SINGLE_OUTPUT_TYPE, TRANSACTION_PAYLOAD_TYPE, UnitsHelper } from "@iota/iota.js";
+import { TAGGED_DATA_PAYLOAD_TYPE, MILESTONE_PAYLOAD_TYPE, BASIC_OUTPUT_TYPE, TRANSACTION_PAYLOAD_TYPE } from "@iota/iota.js";
+import { Converter } from "@iota/util.js";
 import classNames from "classnames";
 import React, { ReactNode } from "react";
 import { RouteComponentProps } from "react-router-dom";
@@ -9,7 +10,7 @@ import { ReactComponent as PlayIcon } from "../../assets/play.svg";
 import { ServiceFactory } from "../../factories/serviceFactory";
 import { IVisualizerCounts } from "../../models/visualizer/IVisualizerCounts";
 import { IVisualizerVertex } from "../../models/visualizer/IVisualizerVertex";
-import { IMpsMetrics } from "../../models/websocket/IMpsMetrics";
+import { IBpsMetrics } from "../../models/websocket/IBpsMetrics";
 import { WebSocketTopic } from "../../models/websocket/webSocketTopic";
 import { EventAggregator } from "../../services/eventAggregator";
 import { MetricsService } from "../../services/metricsService";
@@ -33,7 +34,8 @@ class Visualizer extends AsyncComponent<RouteComponentProps, VisualizerState> {
         Unsolid: 0x8FE6FAFF,
         Referenced: 0x61E884FF,
         Conflicting: 0xFF8B5CFF,
-        Milestone: 0x666AF6FF,
+        Transaction: 0xC061E8FF,
+        Milestone: 0xD92121FF,
         Tip: 0xFFCA62FF,
         Unknown: 0x9AADCEFF
     };
@@ -97,9 +99,9 @@ class Visualizer extends AsyncComponent<RouteComponentProps, VisualizerState> {
     private readonly _themeService: ThemeService;
 
     /**
-     * The mps metrics subscription id.
+     * The bps metrics subscription id.
      */
-    private _mpsMetricsSubscription?: string;
+    private _bpsMetricsSubscription?: string;
 
     /**
      * The resize method
@@ -126,10 +128,11 @@ class Visualizer extends AsyncComponent<RouteComponentProps, VisualizerState> {
         this._themeService = ServiceFactory.get<ThemeService>("theme");
 
         this.state = {
-            mps: "-",
+            bps: "-",
             total: "-",
             tips: "-",
             referenced: "-",
+            transactions: "-",
             conflicting: "-",
             solid: "-",
             isActive: true,
@@ -161,6 +164,9 @@ class Visualizer extends AsyncComponent<RouteComponentProps, VisualizerState> {
                         referenced: counts.total > 0
                             ? `${(counts.referenced / counts.total * 100).toFixed(2)}%`
                             : "-",
+                        transactions: counts.total > 0
+                            ? `${(counts.transactions / counts.total * 100).toFixed(2)}%`
+                            : "-",
                         conflicting: counts.total > 0
                             ? `${(counts.conflicting / counts.total * 100).toFixed(2)}%`
                             : "-",
@@ -171,10 +177,10 @@ class Visualizer extends AsyncComponent<RouteComponentProps, VisualizerState> {
             (referencedId, excludedIds, counts) => this.referenceVertex(referencedId, excludedIds, counts)
         );
 
-        this._mpsMetricsSubscription = this._metricsService.subscribe<IMpsMetrics>(
-            WebSocketTopic.MPSMetrics, data => {
+        this._bpsMetricsSubscription = this._metricsService.subscribe<IBpsMetrics>(
+            WebSocketTopic.BPSMetrics, data => {
                 if (data && this.state.isActive) {
-                    this.setState({ mps: data.new.toString() });
+                    this.setState({ bps: data.new.toString() });
                 }
             });
 
@@ -191,14 +197,28 @@ class Visualizer extends AsyncComponent<RouteComponentProps, VisualizerState> {
     public componentWillUnmount(): void {
         super.componentWillUnmount();
 
-        if (this._mpsMetricsSubscription) {
-            this._metricsService.unsubscribe(this._mpsMetricsSubscription);
-            this._mpsMetricsSubscription = undefined;
+        if (this._bpsMetricsSubscription) {
+            this._metricsService.unsubscribe(this._bpsMetricsSubscription);
+            this._bpsMetricsSubscription = undefined;
         }
 
         this._vizualizerService.unsubscribe();
 
         EventAggregator.unsubscribe("theme", "visualizer");
+
+        // This is a workaround for an issue in Safari
+        // https://github.com/WebKit/WebKit/pull/1693
+        // https://pqina.nl/blog/total-canvas-memory-use-exceeds-the-maximum-limit/
+        if (this._graphElement) {
+            const canvas = this._graphElement.children[0] as HTMLCanvasElement;
+            canvas.width = 1;
+            canvas.height = 1;
+            const ctx = canvas.getContext("2d");
+            ctx?.clearRect(0, 0, 1, 1);
+        }
+
+        this._graph?.clear();
+        this._renderer?.dispose();
 
         this._graph = undefined;
         this._graphics = undefined;
@@ -232,16 +252,16 @@ class Visualizer extends AsyncComponent<RouteComponentProps, VisualizerState> {
                 <div className="stats-panel-container">
                     <div className="card stats-panel">
                         <div className="card--label">
-                            Messages
+                            Blocks
                         </div>
                         <div className="card--value">
                             {this.state.total}
                         </div>
                         <div className="card--label">
-                            MPS
+                            BPS
                         </div>
                         <div className="card--value">
-                            {this.state.mps}
+                            {this.state.bps}
                         </div>
                         <div className="card--label">
                             Tips
@@ -254,6 +274,12 @@ class Visualizer extends AsyncComponent<RouteComponentProps, VisualizerState> {
                         </div>
                         <div className="card--value">
                             {this.state.referenced}
+                        </div>
+                        <div className="card--label">
+                            Transactions
+                        </div>
+                        <div className="card--value">
+                            {this.state.transactions}
                         </div>
                         <div className="card--label">
                             Conflicting
@@ -282,6 +308,10 @@ class Visualizer extends AsyncComponent<RouteComponentProps, VisualizerState> {
                         <div className="key-panel-item">
                             <div className="key-marker vertex-state--referenced" />
                             <div className="key-label">Referenced</div>
+                        </div>
+                        <div className="key-panel-item">
+                            <div className="key-marker vertex-state--transaction" />
+                            <div className="key-label">Transaction</div>
                         </div>
                         <div className="key-panel-item">
                             <div className="key-marker vertex-state--conflicting" />
@@ -329,11 +359,11 @@ class Visualizer extends AsyncComponent<RouteComponentProps, VisualizerState> {
                                 {this.state.selected.vertex.fullId && (
                                     <React.Fragment>
                                         <div className="card--label">
-                                            Message Id
+                                            Block Id
                                         </div>
                                         <div className="card--value">
                                             <a
-                                                href={this.calculateMessageLink(this.state.selected.vertex)}
+                                                href={this.calculateBlockLink(this.state.selected.vertex)}
                                                 target="_blank"
                                                 rel="noopener noreferrer"
                                             >
@@ -349,7 +379,10 @@ class Visualizer extends AsyncComponent<RouteComponentProps, VisualizerState> {
                                                 Total
                                             </div>
                                             <div className="card--value">
-                                                {UnitsHelper.formatBest(this.calculateTotal())}
+                                                {FormatHelper.getInstance().amount(
+                                                    Number(this.calculateTotal()),
+                                                    false
+                                                )}
                                             </div>
                                         </React.Fragment>
                                     )}
@@ -371,31 +404,20 @@ class Visualizer extends AsyncComponent<RouteComponentProps, VisualizerState> {
                                         </React.Fragment>
                                     )}
                                 {this.state.selected.payload &&
-                                    this.state.selected.payload.type === INDEXATION_PAYLOAD_TYPE && (
+                                    this.state.selected.payload.type === TAGGED_DATA_PAYLOAD_TYPE &&
+                                    this.state.selected.payload.tag && (
                                         <React.Fragment>
                                             <div className="card--label">
-                                                Index UTF8
+                                                Tag UTF8
                                             </div>
                                             <div className="card--value">
-                                                <a
-                                                    href={this.calculateIndexedLink(this.state.selected.payload.index)}
-                                                    target="_blank"
-                                                    rel="noopener noreferrer"
-                                                >
-                                                    {Converter.hexToUtf8(this.state.selected.payload.index)}
-                                                </a>
+                                                {Converter.hexToUtf8(this.state.selected.payload.tag)}
                                             </div>
                                             <div className="card--label">
-                                                Index Hex
+                                                Tag Hex
                                             </div>
                                             <div className="card--value">
-                                                <a
-                                                    href={this.calculateIndexedLink(this.state.selected.payload.index)}
-                                                    target="_blank"
-                                                    rel="noopener noreferrer"
-                                                >
-                                                    {this.state.selected.payload.index}
-                                                </a>
+                                                {this.state.selected.payload.tag}
                                             </div>
                                         </React.Fragment>
                                     )}
@@ -450,7 +472,7 @@ class Visualizer extends AsyncComponent<RouteComponentProps, VisualizerState> {
             events.dblClick(node => {
                 this.selectNode();
                 window.open(
-                    this.calculateMessageLink(node.data),
+                    this.calculateBlockLink(node.data),
                     "_blank"
                 );
             });
@@ -462,7 +484,7 @@ class Visualizer extends AsyncComponent<RouteComponentProps, VisualizerState> {
                         this._enteredVertexId = undefined;
                     }
                     if (node) {
-                        this._enteredVertexId = node.data?.shortId ?? node.id.slice(0, 7);
+                        this._enteredVertexId = node.data?.shortId ?? node.id.slice(0, 10);
                         this.connectedLinkStyle(this._enteredVertexId, true);
                     }
                 }
@@ -549,6 +571,9 @@ class Visualizer extends AsyncComponent<RouteComponentProps, VisualizerState> {
             return "Conflicting";
         }
         if (vertex.isReferenced) {
+            if (vertex.isTransaction) {
+                return "Transaction";
+            }
             return "Referenced";
         }
         if (vertex.isSolid) {
@@ -660,7 +685,7 @@ class Visualizer extends AsyncComponent<RouteComponentProps, VisualizerState> {
 
             while (seenNodes.length > pointer) {
                 const node = seenNodes[pointer++];
-                const nodeId = node.data?.shortId ?? node.id.slice(0, 7);
+                const nodeId = node.data?.shortId ?? node.id.slice(0, 10);
 
                 if (nodeCallback?.(nodeId)) {
                     continue;
@@ -761,8 +786,8 @@ class Visualizer extends AsyncComponent<RouteComponentProps, VisualizerState> {
                                 payloadTitle = " - Transaction";
                             } else if (payload.type === MILESTONE_PAYLOAD_TYPE) {
                                 payloadTitle = "";
-                            } else if (payload.type === INDEXATION_PAYLOAD_TYPE) {
-                                payloadTitle = " - Indexation";
+                            } else if (payload.type === TAGGED_DATA_PAYLOAD_TYPE) {
+                                payloadTitle = " - Tagged data";
                             }
                         } else if (node.data.isMilestone) {
                             payloadTitle = " - Checkpoint";
@@ -784,23 +809,15 @@ class Visualizer extends AsyncComponent<RouteComponentProps, VisualizerState> {
     }
 
     /**
-     * Calculate the link for the message.
+     * Calculate the link for the block.
      * @param vertex The vertex id.
-     * @returns The url for the message.
+     * @returns The url for the block.
      */
-    private calculateMessageLink(vertex?: IVisualizerVertex): string {
+    private calculateBlockLink(vertex?: IVisualizerVertex): string {
         return vertex?.fullId
-            ? `${window.location.protocol}//${window.location.host}/explorer/message/${vertex.fullId}`
+            ? `${window.location.protocol}//${window.location.host}` +
+              `${process.env.PUBLIC_URL}/explorer/block/${vertex.fullId}`
             : "";
-    }
-
-    /**
-     * Calculate the link for the index.
-     * @param index The payload index.
-     * @returns The url for the index.
-     */
-    private calculateIndexedLink(index: string): string {
-        return `${window.location.protocol}//${window.location.host}/explorer/indexed/${index}`;
     }
 
     /**
@@ -875,8 +892,8 @@ class Visualizer extends AsyncComponent<RouteComponentProps, VisualizerState> {
 
         if (this.state.selected?.payload?.type === TRANSACTION_PAYLOAD_TYPE) {
             for (const output of this.state.selected.payload.essence.outputs) {
-                if (output.type === SIG_LOCKED_SINGLE_OUTPUT_TYPE) {
-                    total += output.amount;
+                if (output.type === BASIC_OUTPUT_TYPE) {
+                    total += Number(output.amount);
                 }
             }
         }
