@@ -1,16 +1,16 @@
-import { IIndexationPayload, IMilestonePayload, ITransactionPayload } from "@iota/iota.js";
+import { ITaggedDataPayload, IMilestonePayload, ITransactionPayload } from "@iota/iota.js";
 import { IVisualizerVertex } from "../../models/visualizer/IVisualizerVertex";
 
 export interface VisualizerState {
     /**
-     * Messages.
+     * Blocks.
      */
     total: string;
 
     /**
-     * Messages per second.
+     * Blocks per second.
      */
-    mps: string;
+    bps: string;
 
     /**
      * Tips.
@@ -21,6 +21,11 @@ export interface VisualizerState {
      * Referenced.
      */
     referenced: string;
+
+    /**
+     * Transactions.
+     */
+    transactions: string;
 
     /**
      * Conflicting.
@@ -59,7 +64,7 @@ export interface VisualizerState {
         /**
          * Select item payload.
          */
-        payload?: ITransactionPayload | IIndexationPayload | IMilestonePayload;
+        payload?: ITransactionPayload | ITaggedDataPayload | IMilestonePayload;
     };
 
     /**
