@@ -1,3 +1,4 @@
+import { blockIdFromMilestonePayload } from "@iota/iota.js";
 import React, { ReactNode } from "react";
 import { Link, RouteComponentProps } from "react-router-dom";
 import { ReactComponent as ChevronLeftIcon } from "../../../assets/chevron-left.svg";
@@ -7,15 +8,15 @@ import { TangleService } from "../../../services/tangleService";
 import { ClipboardHelper } from "../../../utils/clipboardHelper";
 import { FormatHelper } from "../../../utils/formatHelper";
 import AsyncComponent from "../../components/layout/AsyncComponent";
-import MessageButton from "../../components/layout/MessageButton";
+import BlockButton from "../../components/layout/BlockButton";
 import "./Milestone.scss";
-import { MilestoneRouteProps } from "./MilestoneRouteProps";
+import { MilestoneProps } from "./MilestoneProps";
 import { MilestoneState } from "./MilestoneState";
 
 /**
  * Component which will show the milestone page.
  */
-class Milestone extends AsyncComponent<RouteComponentProps<MilestoneRouteProps>, MilestoneState> {
+class Milestone extends AsyncComponent<RouteComponentProps<MilestoneProps>, MilestoneState> {
     /**
      * Service for tangle requests.
      */
@@ -25,7 +26,7 @@ class Milestone extends AsyncComponent<RouteComponentProps<MilestoneRouteProps>,
      * Create a new instance of Milestone.
      * @param props The props.
      */
-    constructor(props: RouteComponentProps<MilestoneRouteProps>) {
+    constructor(props: RouteComponentProps<MilestoneProps>) {
         super(props);
 
         this._tangleService = ServiceFactory.get<TangleService>("tangle");
@@ -71,21 +72,21 @@ class Milestone extends AsyncComponent<RouteComponentProps<MilestoneRouteProps>,
                             {this.state.milestone?.index}
                         </div>
                         <div className="card--label">
-                            Message Id
+                            Block Id
                         </div>
                         <div className="card--value card--value__mono row">
                             <span className="margin-r-t">
                                 <Link
-                                    to={`/explorer/message/${this.state.milestone?.messageId}`}
+                                    to={`/explorer/block/${this.state.blockId}`}
                                     className="info-box--title linked"
                                 >
-                                    {this.state.milestone?.messageId}
+                                    {this.state.blockId}
                                 </Link>
 
                             </span>
-                            <MessageButton
+                            <BlockButton
                                 onClick={() => ClipboardHelper.copy(
-                                    this.state.milestone?.messageId
+                                    this.state.blockId
                                 )}
                                 buttonType="copy"
                                 labelPosition="top"
@@ -141,6 +142,19 @@ class Milestone extends AsyncComponent<RouteComponentProps<MilestoneRouteProps>,
         const result = await this._tangleService.milestoneDetails(Number.parseInt(index, 10));
 
         if (result) {
+            try {
+                const tangleService = ServiceFactory.get<TangleService>("tangle");
+                const info = await tangleService.info();
+
+                this.setState({
+                    blockId: blockIdFromMilestonePayload(info.protocol.version, result)
+                });
+            } catch (error) {
+                if (error instanceof Error) {
+                    console.log(error.message);
+                }
+            }
+
             this.setState({
                 milestone: result
             }, async () => this.checkForAdjacentMilestones());
