@@ -1,10 +1,15 @@
-import { IMilestoneResponse } from "@iota/iota.js";
+import { IMilestonePayload } from "@iota/iota.js";
 
 export interface MilestoneState {
     /**
      * Milestone.
      */
-    milestone?: IMilestoneResponse;
+    milestone?: IMilestonePayload;
+
+    /**
+     * The block id of the milestone.
+     */
+    blockId?: string;
 
     /**
      * The previous milestone is available.
