@@ -23,6 +23,13 @@ export interface IParticipationEventStatus {
      * The SHA256 checksum of all the question and answer status or the staking amount and
      * rewards calculated for this MilestoneIndex.
      */
-     checksum: string;
+    checksum: string;
+
+    /**
+     * The error message.
+     */
+    error?: {
+        message: string;
+    };
 }
 
