@@ -1,5 +1,5 @@
 /* eslint-disable camelcase */
 export interface IConfirmedInfo {
-    id: string;
-    excluded_ids?: string[];
+    ids: string[];
+    excludedIds?: string[];
 }
