@@ -1,8 +1,6 @@
 /* eslint-disable camelcase */
 export interface IConfirmedMsMetrics {
-    ms_index: number;
-    mps: number;
-    rmps: number;
-    referenced_rate: number;
-    time_since_last_ms: number;
+    blocksPerSecond: number;
+    referencedBlocksPerSecond: number;
+    referencedRate: number;
 }
