@@ -1,7 +1,6 @@
 /* eslint-disable camelcase */
 export interface IPublicNodeStatus {
-    snapshot_index: number;
-    pruning_index: number;
-    is_healthy: boolean;
-    is_synced: boolean;
+    pruningIndex: number;
+    isHealthy: boolean;
+    isSynced: boolean;
 }
