@@ -1,5 +1,5 @@
 /* eslint-disable camelcase */
 export interface ITipInfo {
     id: string;
-    is_tip: boolean;
+    isTip: boolean;
 }
