@@ -1,8 +1,12 @@
-import { Bech32Helper, ClientError, Converter, IClient, IIndexationPayload, IMessageMetadata, IMilestonePayload, IMilestoneResponse, INodeInfo, IOutputResponse, ITransactionPayload, SingleNodeClient } from "@iota/iota.js";
+import { Bech32Helper, ClientError, IClient, ITaggedDataPayload, IBlockMetadata, IMilestonePayload, IRoutesResponse, INodeInfo, IOutputResponse, ITransactionPayload, SingleNodeClient, IndexerPluginClient, ALIAS_ADDRESS_TYPE, NFT_ADDRESS_TYPE } from "@iota/iota.js";
 import { ServiceFactory } from "../factories/serviceFactory";
+import { IAssociatedOutput } from "../models/tangle/IAssociatedOutputsResponse";
+import { ISearchRequest } from "../models/tangle/ISearchRequest";
 import { ISearchResponse } from "../models/tangle/ISearchResponse";
+import { Bech32AddressHelper } from "../utils/bech32AddressHelper";
+import { OutputsHelper } from "../utils/outputsHelper";
+import { SearchQuery, SearchQueryBuilder } from "../utils/searchQueryBuilder";
 import { AuthService } from "./authService";
-
 /**
  * Service to handle api requests.
  */
@@ -25,6 +29,16 @@ export class TangleService {
     }
 
     /**
+     * Get the routes the node exposes.
+     * @returns The routes.
+     */
+     public async routes(): Promise<IRoutesResponse> {
+        const client = this.buildClient();
+        const routes = await client.routes();
+        return routes;
+    }
+
+    /**
      * Get the node info.
      * @returns The node info.
      */
@@ -35,80 +49,29 @@ export class TangleService {
     }
 
     /**
-     * Query the node.
-     * @param query The query to search for.
-     * @returns The response data.
+     * Find item on the stardust network.
+     * @param request The earch query
+     * @returns The item found.
      */
-    public async search(query: string): Promise<ISearchResponse & { unavailable?: boolean }> {
-        const queryLower = query.toLowerCase();
+    public async search(request: ISearchRequest): Promise<ISearchResponse & { unavailable?: boolean }> {
         const client = this.buildClient();
+        const indexerPlugin = new IndexerPluginClient(client);
 
-        try {
-            // If the query is an integer then lookup a milestone
-            if (/^\d+$/.test(query)) {
-                const milestone = await client.milestone(Number.parseInt(query, 10));
+        if (!this._nodeInfo) {
+            await this.info();
+        }
+        const bech32HRP = this._nodeInfo ? this._nodeInfo.protocol.bech32Hrp : Bech32Helper.BECH32_DEFAULT_HRP_MAIN;
+        const searchQuery: SearchQuery = new SearchQueryBuilder(request.query, bech32HRP).build();
+
+        if (searchQuery.milestoneIndex) {
+            try {
+                const milestone = await client.milestoneByIndex(searchQuery.milestoneIndex);
 
                 return {
                     milestone
                 };
-            }
-        } catch (err) {
-            if (this.checkForUnavailable(err)) {
-                return {
-                    unavailable: true
-                };
-            }
-        }
-
-        try {
-            if (!this._nodeInfo) {
-                await this.info();
-            }
-            if (this._nodeInfo && Bech32Helper.matches(queryLower, this._nodeInfo.bech32HRP)) {
-                const address = await client.address(queryLower);
-
-                if (address) {
-                    const addressOutputs = await client.addressOutputs(queryLower);
-
-                    return {
-                        address,
-                        addressOutputIds: addressOutputs.outputIds
-                    };
-                }
-            }
-        } catch (err) {
-            if (this.checkForUnavailable(err)) {
-                return {
-                    unavailable: true
-                };
-            }
-        }
-
-        // If the query is 64 bytes hex, try and look for a message
-        if (Converter.isHex(queryLower) && queryLower.length === 64) {
-            try {
-                const message = await client.message(queryLower);
-
-                return {
-                    message
-                };
             } catch (err) {
-                if (this.checkForUnavailable(err)) {
-                    return {
-                        unavailable: true
-                    };
-                }
-            }
-
-            // If the query is 64 bytes hex, try and look for a transaction included message
-            try {
-                const message = await client.transactionIncludedMessage(queryLower);
-
-                return {
-                    message
-                };
-            } catch (err) {
-                if (this.checkForUnavailable(err)) {
+                if (err instanceof ClientError && this.checkForUnavailable(err)) {
                     return {
                         unavailable: true
                     };
@@ -116,79 +79,157 @@ export class TangleService {
             }
         }
 
-        try {
-            // If the query is 68 bytes hex, try and look for an output
-            if (Converter.isHex(queryLower) && queryLower.length === 68) {
-                const output = await client.output(queryLower);
+        if (searchQuery.milestoneId) {
+            try {
+                const milestone = await client.milestoneById(searchQuery.milestoneId);
 
                 return {
-                    output
+                    milestone
                 };
-            }
-        } catch (err) {
-            if (this.checkForUnavailable(err)) {
-                return {
-                    unavailable: true
-                };
+            } catch (err) {
+                if (err instanceof ClientError && this.checkForUnavailable(err)) {
+                    return {
+                        unavailable: true
+                    };
+                }
             }
         }
 
-        try {
-            if (Converter.isHex(queryLower) && queryLower.length === 64) {
-                // We have 64 characters hex so could possible be a raw ed25519 address
-                const address = await client.addressEd25519(queryLower);
+        if (searchQuery.blockIdOrTransactionId) {
+            try {
+                const block = await client.block(searchQuery.blockIdOrTransactionId);
 
-                const addressOutputs = await client.addressEd25519Outputs(queryLower);
+                if (Object.keys(block).length > 0) {
+                    return {
+                        block
+                    };
+                }
+            } catch (err) {
+                if (err instanceof ClientError && this.checkForUnavailable(err)) {
+                    return {
+                        unavailable: true
+                    };
+                }
+            }
 
-                if (addressOutputs.count > 0) {
+            try {
+                const block = await client.transactionIncludedBlock(searchQuery.blockIdOrTransactionId);
+
+                if (Object.keys(block).length > 0) {
+                    return {
+                        block
+                    };
+                }
+            } catch (err) {
+                if (err instanceof ClientError && this.checkForUnavailable(err)) {
+                    return {
+                        unavailable: true
+                    };
+                }
+            }
+        }
+
+        if (searchQuery.output) {
+            try {
+                return {
+                    outputId: searchQuery.output,
+                    output: await client.output(searchQuery.output)
+                };
+            } catch (err) {
+                if (err instanceof ClientError && this.checkForUnavailable(err)) {
+                    return {
+                        unavailable: true
+                    };
+                }
+            }
+        }
+
+        if (searchQuery.aliasId) {
+            try {
+                const aliasOutputs = await indexerPlugin.alias(searchQuery.aliasId);
+                if (aliasOutputs.items.length > 0) {
+                    const address = Bech32AddressHelper
+                                        .buildAddress(searchQuery.aliasId, bech32HRP, ALIAS_ADDRESS_TYPE);
                     return {
                         address,
-                        addressOutputIds: addressOutputs.outputIds
+                        addressOutputIds: aliasOutputs.items
                     };
                 }
-            }
-        } catch (err) {
-            if (this.checkForUnavailable(err)) {
-                return {
-                    unavailable: true
-                };
+            } catch (err) {
+                if (err instanceof ClientError && this.checkForUnavailable(err)) {
+                    return {
+                        unavailable: true
+                    };
+                }
             }
         }
 
-        try {
-            if (query.length > 0) {
-                let messages;
-                let indexMessageType: "utf8" | "hex" | undefined;
-
-                // If the query is between 2 and 128 hex chars assume hex encoded bytes
-                if (query.length >= 2 && query.length <= 128 && Converter.isHex(queryLower)) {
-                    messages = await client.messagesFind(Converter.hexToBytes(queryLower));
-
-                    if (messages.count > 0) {
-                        indexMessageType = "hex";
-                    }
-                }
-
-                // If not already found and query less than 64 bytes assume its UTF8
-                if (!indexMessageType && query.length <= 64) {
-                    messages = await client.messagesFind(query);
-                    if (messages.count > 0) {
-                        indexMessageType = "utf8";
-                    }
-                }
-
-                if (messages && messages.count > 0) {
+        if (searchQuery.nftId) {
+            try {
+                const nftOutputs = await indexerPlugin.nft(searchQuery.nftId);
+                if (nftOutputs.items.length > 0) {
+                    const address = Bech32AddressHelper
+                                    .buildAddress(searchQuery.nftId, bech32HRP, NFT_ADDRESS_TYPE);
                     return {
-                        indexMessageIds: messages.messageIds,
-                        indexMessageType
+                        address,
+                        addressOutputIds: nftOutputs.items
+                    };
+                }
+            } catch (err) {
+                if (err instanceof ClientError && this.checkForUnavailable(err)) {
+                    return {
+                        unavailable: true
                     };
                 }
             }
-        } catch (err) {
-            if (this.checkForUnavailable(err)) {
+        }
+
+        if (searchQuery.foundryId) {
+            try {
+                const foundryOutputs = await indexerPlugin.foundry(searchQuery.foundryId);
+                if (foundryOutputs.items.length > 0) {
+                    return {
+                        outputId: foundryOutputs.items[0]
+                    };
+                }
+            } catch (err) {
+                if (err instanceof ClientError && this.checkForUnavailable(err)) {
+                    return {
+                        unavailable: true
+                    };
+                }
+            }
+        }
+
+        if (searchQuery.tag) {
+            try {
+                const taggedOutputs = await indexerPlugin.basicOutputs({ tagHex: searchQuery.tag });
+
+                if (taggedOutputs.items.length > 0) {
+                    return {
+                        outputs: taggedOutputs.items
+                    };
+                }
+            } catch (err) {
+                if (err instanceof ClientError && this.checkForUnavailable(err)) {
+                    return {
+                        unavailable: true
+                    };
+                }
+            }
+        }
+
+        if (searchQuery.address?.hexNoPrefix && searchQuery.address?.hexNoPrefix.length === 64) {
+            try {
                 return {
-                    unavailable: true
+                    address: { ...searchQuery.address }
                 };
+            } catch (err) {
+                if (err instanceof ClientError && this.checkForUnavailable(err)) {
+                    return {
+                        unavailable: true
+                    };
+                }
             }
         }
 
@@ -196,18 +237,18 @@ export class TangleService {
     }
 
     /**
-     * Get the message payload.
-     * @param messageId The message to get.
+     * Get the block payload.
+     * @param blockId The block to get.
      * @returns The response data.
      */
     public async payload(
-        messageId: string): Promise<ITransactionPayload | IIndexationPayload | IMilestonePayload | undefined> {
+        blockId: string): Promise<ITransactionPayload | ITaggedDataPayload | IMilestonePayload | undefined> {
         try {
             const client = this.buildClient();
 
-            const message = await client.message(messageId);
+            const block = await client.block(blockId);
 
-            return message?.payload;
+            return block?.payload;
         } catch {}
     }
 
@@ -226,43 +267,116 @@ export class TangleService {
     }
 
     /**
+     * Get the nft output details.
+     * @param nftId The nft id to get the details for.
+     * @returns The nft output response.
+     */
+     public async nftDetails(
+        nftId: string): Promise<IOutputResponse & { outputId: string } | undefined> {
+        try {
+            const client = this.buildClient();
+            const indexerPlugin = new IndexerPluginClient(client);
+
+            const nftOutputs = await indexerPlugin.nft(nftId);
+            if (nftOutputs.items.length > 0) {
+                return { ...await client.output(nftOutputs.items[0]), outputId: nftOutputs.items[0] };
+            }
+        } catch {}
+    }
+
+    /**
+     * Get the alias output details.
+     * @param aliasId The alias id to get the details for.
+     * @returns The alias output response.
+     */
+     public async aliasDetails(
+        aliasId: string): Promise<IOutputResponse & { outputId: string } | undefined> {
+        try {
+            const client = this.buildClient();
+            const indexerPlugin = new IndexerPluginClient(client);
+
+            const aliasOutputs = await indexerPlugin.alias(aliasId);
+            if (aliasOutputs.items.length > 0) {
+                return { ...await client.output(aliasOutputs.items[0]), outputId: aliasOutputs.items[0] };
+            }
+        } catch {}
+    }
+
+    /**
+     * Get the output details.
+     * @param addressBech32 The address to get the outputs for.
+     * @returns The associated outputs.
+     */
+    public async getOutputsForAddress(
+        addressBech32: string): Promise<IAssociatedOutput[]> {
+        let outputs: IAssociatedOutput[] = [];
+
+        try {
+            const client = this.buildClient();
+            const helper = new OutputsHelper(addressBech32, client);
+            await helper.fetchAssociatedOutputs();
+            outputs = helper.associatedOutputs;
+        } catch {}
+
+        return outputs;
+    }
+
+    /**
+     * Get the output details.
+     * @param tag The tag to get the outputs for.
+     * @returns The associated outputs.
+     */
+    public async getOutputsByTag(
+        tag: string): Promise<IAssociatedOutput[]> {
+        let outputs: IAssociatedOutput[] = [];
+
+        try {
+            const client = this.buildClient();
+            const helper = new OutputsHelper(tag, client);
+            await helper.fetchOutputsByTag();
+            outputs = helper.associatedOutputs;
+        } catch {}
+
+        return outputs;
+    }
+
+    /**
      * Get the miletsone details.
      * @param milestoneIndex The miletsone to get the details for.
      * @returns The details response.
      */
     public async milestoneDetails(
-        milestoneIndex: number): Promise<IMilestoneResponse | undefined> {
+        milestoneIndex: number): Promise<IMilestonePayload | undefined> {
         try {
             const client = this.buildClient();
 
-            return await client.milestone(milestoneIndex);
+            return await client.milestoneByIndex(milestoneIndex);
         } catch {}
     }
 
     /**
-     * Get the message metadata.
-     * @param messageId The message if to get the metadata for.
+     * Get the block metadata.
+     * @param blockId The block if to get the metadata for.
      * @returns The details response.
      */
-    public async messageDetails(messageId: string): Promise<{
-        metadata?: IMessageMetadata;
-        childrenIds?: string[];
+    public async blockDetails(blockId: string): Promise<{
+        metadata?: IBlockMetadata;
         unavailable?: boolean;
     } | undefined> {
         try {
             const client = this.buildClient();
 
-            const metadata = await client.messageMetadata(messageId);
-            const children = await client.messageChildren(messageId);
+            const metadata = await client.blockMetadata(blockId);
 
             return {
-                metadata,
-                childrenIds: children ? children.childrenMessageIds : undefined
+                metadata
             };
         } catch (err) {
-            return {
-                unavailable: this.checkForUnavailable(err)
-            };
+            if (err instanceof ClientError && this.checkForUnavailable(err)) {
+                return {
+                    unavailable: true
+                };
+            }
         }
     }
 
@@ -287,6 +401,7 @@ export class TangleService {
         await client.peerDelete(peerId);
     }
 
+
     /**
      * Build a client with auth header.
      * @returns The client.
@@ -307,7 +422,7 @@ export class TangleService {
         return new SingleNodeClient(
             `${window.location.protocol}//${window.location.host}`,
             {
-                basePath: "/api/v1/",
+                basePath: "/dashboard/api/",
                 headers
             });
     }
