@@ -1,4 +1,5 @@
-import { Bech32Helper, Converter, ED25519_ADDRESS_TYPE } from "@iota/iota.js";
+import { Bech32Helper, ALIAS_ADDRESS_TYPE, ED25519_ADDRESS_TYPE, NFT_ADDRESS_TYPE } from "@iota/iota.js";
+import { Converter, HexHelper } from "@iota/util.js";
 import { IBech32AddressDetails } from "../models/IBech32AddressDetails";
 
 export class Bech32AddressHelper {
@@ -6,11 +7,13 @@ export class Bech32AddressHelper {
      * Build the address details.
      * @param address The address to source the data from.
      * @param hrp The human readable part of the address.
+     * @param addressType The address type
      * @returns The parts of the address.
      */
-    public static buildAddress(address: string, hrp: string): IBech32AddressDetails {
+    public static buildAddress(address: string, hrp: string, addressType?: number): IBech32AddressDetails {
         let bech32;
         let hex;
+        let hexNoPrefix;
         let type;
 
         if (Bech32Helper.matches(address, hrp)) {
@@ -19,21 +22,23 @@ export class Bech32AddressHelper {
                 if (result) {
                     bech32 = address;
                     type = result.addressType;
-                    hex = Converter.bytesToHex(result.addressBytes);
+                    hex = Converter.bytesToHex(result.addressBytes, true);
+                    hexNoPrefix = HexHelper.stripPrefix(hex);
                 }
             } catch {}
         }
 
         if (!bech32) {
-            // We assume this is hex and ed25519 for now
-            hex = address;
-            type = ED25519_ADDRESS_TYPE;
-            bech32 = Bech32Helper.toBech32(ED25519_ADDRESS_TYPE, Converter.hexToBytes(hex), hrp);
+            hex = HexHelper.addPrefix(address);
+            hexNoPrefix = HexHelper.stripPrefix(address);
+            type = addressType !== undefined ? addressType : ED25519_ADDRESS_TYPE;
+            bech32 = Bech32Helper.toBech32(type, Converter.hexToBytes(hex), hrp);
         }
 
         return {
             bech32,
             hex,
+            hexNoPrefix,
             type,
             typeLabel: Bech32AddressHelper.typeLabel(type)
         };
@@ -47,6 +52,11 @@ export class Bech32AddressHelper {
     public static typeLabel(addressType?: number): string | undefined {
         if (addressType === ED25519_ADDRESS_TYPE) {
             return "Ed25519";
+        } else if (addressType === ALIAS_ADDRESS_TYPE) {
+            return "Alias";
+        } else if (addressType === NFT_ADDRESS_TYPE) {
+            return "Nft";
         }
+            return "Unknown address type";
     }
 }
